Fix for CM-1580
The way ProcessStats.java reads the time_in_state_file is silly, but
minimizes GC. The buffer used is too small on newer devices and
overflows. Enlarge it so we have correct stats.

Also fix the bad assumption in getCpuSpeedTimes that the file will never
be truncated. This can cause system_server to go into an exception
loop.